retrains using grid search parameters

+++ scripts/train_subtype.py
@@ -1,5 +1,6 @@
 from model_setup import *
 import re
+import os
 import sys
 import json
 import argparse
@@ -13,7 +14,7 @@ parser.add_argument(
 args = parser.parse_args()
 
 # only ['skgrid', 'aklimate', 'subscope'] models have GEXP features shared across TCGA, AURORA and METABRIC
-with open('./tools/model_info.json', 'r') as fh:
+with open('./models/model_info.json', 'r') as fh:
     data = json.load(fh)
     if args.featureset == 'all':
         ft_list = []
@@ -36,16 +37,20 @@ df['Labels'] = df['Labels'].replace(label_dict_tcga)
 X_train, X_val, y_train, y_val = train_test_split(
     df[ft_set],
     df['Labels'],
-    test_size=0.15,
+    test_size=0.1,
     stratify=df['Labels'],
     random_state=333
 )
-# init and fit simple scaler
-scaler = MinMaxScaler()
-scaler.fit(X_train)
-# pickle scaler for subsequent use of trained SAGE model
-with open(f'./models/scalers/{args.featureset}_scaler.pickle', 'wb') as file:
-    pickle.dump(scaler, file)
+# check if scaler exists
+if os.path.exists(f'./models/scalers/{args.featureset}_scaler.pickle'):
+    pass
+else:
+    # init and fit simple scaler
+    scaler = MinMaxScaler()
+    scaler.fit(X_train)
+    # pickle scaler for subsequent use of trained SAGE model
+    with open(f'./models/scalers/{args.featureset}_scaler.pickle', 'wb') as file:
+        pickle.dump(scaler, file)
 
 # convert back to joint df for upscaling
 train_df = pd.DataFrame(X_train, columns=ft_set)
@@ -53,8 +58,8 @@ train_df['Labels'] = y_train
 # upsample minor classes to reduce imbalance
 df_0 = train_df.loc[df['Labels'] == 0].copy() # 1x (from 454)
 df_1 = pd.DataFrame(np.repeat(train_df.loc[train_df['Labels'] == 1], 2, axis=0), columns=train_df.columns) # 2x (to 348)
-df_2 = pd.DataFrame(np.repeat(train_df.loc[train_df['Labels'] == 2], 2, axis=0), columns=train_df.columns) # 2x (to 298)
-df_3 = pd.DataFrame(np.repeat(train_df.loc[train_df['Labels'] == 3], 5, axis=0), columns=train_df.columns) # 5x (to 340)
+df_2 = pd.DataFrame(np.repeat(train_df.loc[train_df['Labels'] == 2], 3, axis=0), columns=train_df.columns) # 3x (to 447)
+df_3 = pd.DataFrame(np.repeat(train_df.loc[train_df['Labels'] == 3], 6, axis=0), columns=train_df.columns) # 6x (to 414)
 train_df = pd.concat([df_0, df_1, df_2, df_3])
 X_train = train_df[ft_set]
 y_train = train_df['Labels']
@@ -71,21 +76,22 @@ train_data = CustomTensorDataset(
     np.array(scaled_X_train).astype(np.float32),
     np.array(y_train).astype(np.int64)
 )
-model = FcSAE(in_features=len(ft_set))
-trained_model, history = train_sae(model, train_data, epochs=50)
-# save trained model
-model_path = f'./models/{args.featureset}_sage'
-torch.save(trained_model.state_dict(), model_path)
-
-# calibrate model classifier
-base_model = FcSAE(in_features=len(ft_set))
-base_model.load_state_dict(torch.load(model_path)) # load model from save point
-calib_model = ModelWithTemperature(base_model)
-# tune temperature param using validation set
 val_data = CustomTensorDataset(
     np.array(scaled_X_val).astype(np.float32),
     np.array(y_val).astype(np.int64)
 )
+dims = 10
+model = FcSAE(in_features=len(ft_set), latent_dim=dims)
+## train model
+trained_model, history = train_sae(model, train_data, val_data, epochs=150, lr=3e-4, recon_weight=0.5, task_weight=0.1)
+# save trained model
+model_path = f'./models/{args.featureset}_sage'
+torch.save(trained_model.state_dict(), model_path)
+# calibrate model classifier
+base_model = FcSAE(in_features=len(ft_set), latent_dim=dims)
+base_model.load_state_dict(torch.load(model_path)) # load model from save point
+calib_model = ModelWithTemperature(base_model)
+# tune temperature param using validation set
 val_dl = DataLoader(val_data, batch_size=64, shuffle=False)
 set_temperature(calib_model, val_dl)
 # save tempscaled model
@@ -102,7 +108,7 @@ preds = [item for sublist in preds for item in sublist]
 train_acc = accuracy_score(train_data.targets, preds)
 train_acc = round(train_acc, 4)
 print('\n')
-print(f'Training accuracy: {train_acc}')
+print(f'Last train accuracy: {train_acc}')
 
 # check classification performance
 dl = DataLoader(val_data, batch_size=64, shuffle=False)
@@ -113,4 +119,4 @@ for data, labels in dl:
 preds = [item for sublist in preds for item in sublist]
 val_acc = accuracy_score(val_data.targets, preds)
 val_acc = round(val_acc, 4)
-print(f'Validation accuracy: {val_acc}')
+print(f'Last validation accuracy: {val_acc}')
